Bridges: Integrate Action Bar, delete on edit.

+++ app/bridges/bridges.php
@@ -39,25 +39,45 @@
 	$language = new text;
 	$text = $language->get();
 
-//get the action
-	if (is_array($_POST["bridges"])) {
-		$bridges = $_POST["bridges"];
-		foreach($bridges as $row) {
-			if ($row['action'] == 'delete') {
-				$action = 'delete';
-				break;
-			}
+//get posted data
+	if (is_array($_POST['bridges'])) {
+		$action = $_POST['action'];
+		$bridges = $_POST['bridges'];
+	}
+
+//copy the bridges
+	if (permission_exists('bridge_add')) {
+		if ($action == 'copy' && is_array($bridges) && @sizeof($bridges) != 0) {
+			//copy
+				$obj = new bridges;
+				$obj->copy($bridges);
+			//redirect
+				header('Location: bridges.php');
+				exit;
+		}
+	}
+
+//toggle the bridges
+	if (permission_exists('bridge_edit')) {
+		if ($action == 'toggle' && is_array($bridges) && @sizeof($bridges) != 0) {
+			//toggle
+				$obj = new bridges;
+				$obj->toggle($bridges);
+			//redirect
+				header('Location: bridges.php');
+				exit;
 		}
 	}
 
 //delete the bridges
 	if (permission_exists('bridge_delete')) {
-		if ($action == "delete") {
-			//download
+		if ($action == 'delete' && is_array($bridges) && @sizeof($bridges) != 0) {
+			//delete
 				$obj = new bridges;
 				$obj->delete($bridges);
-			//delete message
-				message::add($text['message-delete']);
+			//redirect
+				header('Location: bridges.php');
+				exit;
 		}
 	}
 
@@ -107,7 +127,7 @@
 
 //get the list
 	$sql = str_replace('count(*)', '*', $sql);
-	$sql .= order_by($order_by, $order);
+	$sql .= order_by($order_by, $order, 'bridge_name', 'asc');
 	$sql .= limit_offset($rows_per_page, $offset);
 	$database = new database;
 	$bridges = $database->select($sql, $parameters, 'all');
@@ -124,56 +144,98 @@
 //include the header
 	require_once "resources/header.php";
 
-//define the checkbox_toggle function
-	echo "<script type=\"text/javascript\">\n";
-	echo "	function checkbox_toggle(item) {\n";
-	echo "		var inputs = document.getElementsByTagName(\"input\");\n";
+//define the checkbox functions
+	echo "<script type='text/javascript'>\n";
+	echo "	function checkbox_toggle() {\n";
+	echo "		var inputs = document.getElementsByTagName('input');\n";
+	echo "		var box_checked = document.getElementById('checkbox_all').checked;\n";
 	echo "		for (var i = 0, max = inputs.length; i < max; i++) {\n";
 	echo "		    if (inputs[i].type === 'checkbox') {\n";
-	echo "		       	if (document.getElementById('checkbox_all').checked == true) {\n";
-	echo "				inputs[i].checked = true;\n";
-	echo "			}\n";
-	echo "				else {\n";
-	echo "					inputs[i].checked = false;\n";
-	echo "				}\n";
+	echo "				inputs[i].checked = box_checked;\n";
 	echo "			}\n";
 	echo "		}\n";
+	echo "		if (box_checked) {\n";
+	echo "			document.getElementById('btn_check_all').style.display = 'none';\n";
+	echo "			document.getElementById('btn_check_none').style.display = '';\n";
+	echo "		}\n";
+	echo "		else {\n";
+	echo "			document.getElementById('btn_check_all').style.display = '';\n";
+	echo "			document.getElementById('btn_check_none').style.display = 'none';\n";
+	echo "		}\n";
 	echo "	}\n";
+	/*
+	echo "	function check_all() {\n";
+	echo "		var inputs = document.getElementsByTagName('input');\n";
+	echo "		for (var i = 0, max = inputs.length; i < max; i++) {\n";
+	echo "		    if (inputs[i].type === 'checkbox') {\n";
+	echo "		       	inputs[i].checked = true;\n";
+	echo "			}\n";
+	echo "		}\n";
+	echo "		document.getElementById('btn_check_all').style.display = 'none';\n";
+	echo "		document.getElementById('btn_check_none').style.display = '';\n";
+	echo "	}\n";
+
+	echo "	function check_none() {\n";
+	echo "		var inputs = document.getElementsByTagName('input');\n";
+	echo "		for (var i = 0, max = inputs.length; i < max; i++) {\n";
+	echo "		    if (inputs[i].type === 'checkbox') {\n";
+	echo "		       	inputs[i].checked = false;\n";
+	echo "			}\n";
+	echo "		}\n";
+	echo "		document.getElementById('btn_check_all').style.display = '';\n";
+	echo "		document.getElementById('btn_check_none').style.display = 'none';\n";
+	echo "	}\n";
+	*/
 	echo "</script>\n";
 
 //show the content
-	echo "<table width='100%' border='0'>\n";
-	echo "	<tr>\n";
-	echo "		<td width='50%' align='left' nowrap='nowrap'><b>".$text['title-bridges']."</b></td>\n";
-	echo "		<form method='get' action=''>\n";
-	echo "			<td width='50%' style='vertical-align: top; text-align: right; white-space: nowrap;'>\n";
-
-	if (permission_exists('bridge_all')) {
-		if ($_GET['show'] == 'all') {
-			echo "		<input type='hidden' name='show' value='all'>";
+	echo "<div class='action_bar'>\n";
+	echo "	<b style='float: left;'>".$text['title-bridges']." (".$num_rows.")</b>\n";
+	if (permission_exists('bridge_add')) {
+		echo "	<a href='bridge_edit.php' alt=\"".$text['button-add']."\"><button type='button' class='btn btn-default'>".$text['button-add']."</button></a>";
+	}
+	if (permission_exists('bridge_add') || permission_exists('bridge_edit') || permission_exists('bridge_delete')) {
+		/*
+		echo "	<button type='button' class='btn btn-default' id='btn_check_all' alt=\"".$text['button-check_all']."\" onclick='check_all();'>".$text['button-check_all']."</button>\n";
+		echo "	<button type='button' class='btn btn-default' id='btn_check_none' alt=\"".$text['button-check_none']."\" onclick='check_none();' style='display: none;'>".$text['button-check_none']."</button>\n";
+		*/
+		if (permission_exists('bridge_add')) {
+			echo "	<button type='button' class='btn btn-default' alt=\"".$text['button-copy']."\" onclick=\"if (confirm('".$text['confirm-copy']."')) { set_action('copy'); submit_form('form_list'); } else { this.blur(); return false; }\">".$text['button-copy']."</button>";
 		}
-		else {
-			echo "		<input type='button' class='btn' value='".$text['button-show_all']."' onclick=\"window.location='bridges.php?show=all';\">\n";
+		if (permission_exists('bridge_edit')) {
+			echo "	<button type='button' class='btn btn-default' alt=\"".$text['button-toggle']."\" onclick=\"if (confirm('".$text['confirm-toggle']."')) { set_action('toggle'); submit_form('form_list'); } else { this.blur(); return false; }\">".$text['button-toggle']."</button>";
+		}
+		if (permission_exists('bridge_delete')) {
+			echo "	<button type='button' class='btn btn-default' alt=\"".$text['button-delete']."\" onclick=\"if (confirm('".$text['confirm-delete']."')) { set_action('delete'); submit_form('form_list'); } else { this.blur(); return false; }\">".$text['button-delete']."</button>";
 		}
 	}
+	echo "	<span style='white-space: nowrap;'>\n";
+	if (permission_exists('bridge_all')) {
+		if ($_GET['show'] == 'all') {
+			echo "	<input type='hidden' name='show' value='all'>";
+		}
+		else {
+			echo "	<a href='?show=all'><button type='button' class='btn btn-default' style='margin-left: 15px;'>".$text['button-show_all']."</button></a>\n";
+		}
+	}
+	echo "	<form id='form_search' class='inline' method='get'>\n";
+	echo "	<input type='text' class='txt' style='width: 150px; margin-left: 15px;' name='search' id='search' value='".escape($search)."' onkeypress='reset_search();'>\n";
+	echo "	<input type='submit' class='btn' id='btn_search' value=\"".$text['button-search']."\" style='".($search != '' ? 'display: none;' : null)."'>\n";
+	echo "	<a href='bridges.php'><input type='button' class='btn' id='btn_reset' value=\"".$text['button-reset']."\" style='".($search == '' ? 'display: none;' : null)."'></a>\n";
+	echo "	</form>\n";
+	echo "	</span>\n";
+	echo "</div>\n";
 
-	echo "				<input type='text' class='txt' style='width: 150px; margin-left: 15px;' name='search' id='search' value='".escape($search)."'>\n";
-	echo "				<input type='submit' class='btn' name='submit' value='".$text['button-search']."'>\n";
-	echo "			</td>\n";
-	echo "		</form>\n";
-	echo "	</tr>\n";
-	echo "	<tr>\n";
-	echo "		<td align='left' colspan='2'>\n";
-	echo "			".$text['title_description-bridge']."<br /><br />\n";
-	echo "		</td>\n";
-	echo "	</tr>\n";
-	echo "</table>\n";
+	echo $text['title_description-bridge']."\n";
+	echo "<br /><br />\n";
 
-	echo "<form method='post' action=''>\n";
+	echo "<form id='form_list' method='post'>\n";
+	echo "<input type='hidden' id='action' name='action' value=''>\n";
+
 	echo "<table class='tr_hover' width='100%' border='0' cellpadding='0' cellspacing='0'>\n";
 	echo "<tr>\n";
-	echo "	<th style='width:30px;'>\n";
-	echo "		<input type='checkbox' name='checkbox_all' id='checkbox_all' value='' onclick=\"checkbox_toggle();\">\n";
+	echo "	<th style='width: 30px;'>\n";
+	echo "		<input type='checkbox' id='checkbox_all' name='checkbox_all' value='' onclick='checkbox_toggle();'>\n";
 	echo "	</th>\n";
 	if ($_GET['show'] == "all" && permission_exists('bridge_all')) {
 		echo th_order_by('domain_name', $text['label-domain'], $order_by, $order);
@@ -181,15 +243,9 @@
 	echo th_order_by('bridge_name', $text['label-bridge_name'], $order_by, $order);
 	echo th_order_by('bridge_destination', $text['label-bridge_destination'], $order_by, $order);
 	echo th_order_by('bridge_enabled', $text['label-bridge_enabled'], $order_by, $order);
-	echo "	<td class='list_control_icons'>";
-	if (permission_exists('bridge_add')) {
-		echo "		<a href='bridge_edit.php' alt='".$text['button-add']."'>$v_link_label_add</a>";
-	}
-	else {
-		echo "&nbsp;\n";
-	}
-	echo "	</td>\n";
-	echo "<tr>\n";
+	echo "	<th>".$text['label-description']."</th>\n";
+	echo "	<td style='width: 1px;'>&nbsp;</td>\n";
+	echo "</tr>\n";
 
 	if (is_array($bridges)) {
 		$x = 0;
@@ -199,42 +255,64 @@
 			}
 			echo "<tr ".$tr_link.">\n";
 			echo "	<td valign='top' class='".$row_style[$c]." tr_link_void' style='align: center; padding: 3px 3px 0px 8px;'>\n";
-			echo "		<input type='checkbox' name=\"bridges[$x][checked]\" id='checkbox_".$x."' value='true' onclick=\"if (!this.checked) { document.getElementById('chk_all_".$x."').checked = false; }\">\n";
-			echo "		<input type='hidden' name=\"bridges[$x][bridge_uuid]\" value='".escape($row['bridge_uuid'])."' />\n";
+			echo "		<input type='checkbox' name='bridges[$x][checked]' id='checkbox_".$x."' value='true' onclick=\"if (!this.checked) { document.getElementById('checkbox_all').checked = false; }\">\n";
+			echo "		<input type='hidden' name='bridges[$x][bridge_uuid]' value='".escape($row['bridge_uuid'])."' />\n";
 			echo "	</td>\n";
 			if ($_GET['show'] == "all" && permission_exists('bridge_all')) {
-				echo "	<td valign='top' class='".$row_style[$c]."'>".escape($_SESSION['domains'][$row['domain_uuid']]['domain_name'])."&nbsp;</td>\n";
+				echo "	<td valign='top' class='".$row_style[$c]."'>".escape($_SESSION['domains'][$row['domain_uuid']]['domain_name'])."</td>\n";
 			}
-			echo "	<td valign='top' class='".$row_style[$c]."'>".escape($row['bridge_name'])."&nbsp;</td>\n";
-			echo "	<td valign='top' class='".$row_style[$c]."'>".escape($row['bridge_destination'])."&nbsp;</td>\n";
-			echo "	<td valign='top' class='".$row_style[$c]."'>".$text['label-'.$row['bridge_enabled']]."&nbsp;</td>\n";
-			echo "	<td class='list_control_icons'>";
+			echo "	<td valign='top' class='".$row_style[$c]."'>";
 			if (permission_exists('bridge_edit')) {
-				echo "<a href='bridge_edit.php?id=".escape($row['bridge_uuid'])."' alt='".$text['button-edit']."'>$v_link_label_edit</a>";
+				echo "	<a ".$tr_link.">".escape($row['bridge_name'])."</a>\n";
 			}
-			if (permission_exists('bridge_delete')) {
-				echo "<button type='submit' class='btn btn-default list_control_icon' name=\"bridges[$x][action]\" alt='".$text['button-delete']."' value='delete'><span class='fas fa-minus'></span></button>";
+			else {
+				echo "	".escape($row['bridge_name'])."\n";
 			}
 			echo "	</td>\n";
+			echo "	<td valign='top' class='".$row_style[$c]."'>".escape($row['bridge_destination'])."</td>\n";
+			echo "	<td valign='top' class='".$row_style[$c]."'>".$text['label-'.$row['bridge_enabled']]."</td>\n";
+			echo "	<td valign='top' class='".$row_style[$c]."'>".escape($row['bridge_description'])."</td>\n";
+			echo "	<td style='width: 0;' class='tr_link_void'>";
+			if (permission_exists('bridge_edit')) {
+				echo "<a href='bridge_edit.php?id=".escape($row['bridge_uuid'])."' alt='".$text['button-edit']."'>".$v_button_icon_edit."</a>";
+			}
+			/*
+			if (permission_exists('bridge_delete')) {
+				echo "<button type='input' class='btn btn-default btn-icon' name='bridges[$x][checked]' value='true' alt='".$text['button-delete']."' onclick=\"if (confirm('".$text['confirm-delete']."')) { set_action('delete'); submit_form('form_list'); } else { this.blur(); return false; }\"><span class='fas fa-minus'></span></button>";
+			}
+			*/
+			echo "	</td>\n";
 			echo "</tr>\n";
+			$c = $c ? 0 : 1;
 			$x++;
-			$c = $c == 1 ? 0 : 1;
-		} //end foreach
+		}
 		unset($sql, $bridges);
-	} //end if results
+	}
 
 	echo "<tr>\n";
 	echo "</table>\n";
-	if (permission_exists('bridge_add')) {
-		echo "<div style='float: right;'>\n";
-		echo "	<a href='bridge_edit.php' alt=\"".$text['button-add']."\">".$v_link_label_add."</a>";
-		echo "</div>\n";
-	}
 	echo "<br />\n";
 	echo "<div align='center'>".$paging_controls."</div>\n";
 
 	echo "<input type='hidden' name='".$token['name']."' value='".$token['hash']."'>\n";
+
 	echo "</form>\n";
+
+//handle form actions
+	echo "<script type='text/javascript'>\n";
+	echo "	function set_action(action) {\n";
+	echo "		document.getElementById('action').value = action;\n";
+	echo "	}\n";
+
+	echo "	function submit_form(form_id) {\n";
+	echo "		document.getElementById(form_id).submit();\n";
+	echo "	}\n";
+
+	echo "	function reset_search() {\n";
+	echo "		document.getElementById('btn_reset').style.display = 'none';\n";
+	echo "		document.getElementById('btn_search').style.display = '';\n";
+	echo "	}\n";
+	echo "</script>\n";
 
 //include the footer
 	require_once "resources/footer.php";
